requirements.txt
fastapi[standard]==0.117.1
uvicorn==0.36.0
httpx==0.28.1
pydantic==2.11.9
python-dotenv==1.1.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
